chore(refactor): refactor types to corresponding entity
